+++ test/cmdlineTests/standard_json_invalid_utf8_in_source_snippet/invalid_utf8_sequence.sol
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: GPL-3.0
+pragma solidity >=0.0;
+
+contract C {
+    // The comment below is encoded in Latin-1 and is not valid UTF-8.
+    uint x = "a"; // résumé
+}